Fix CMakeLists.txt and installation
