+++ Lib/include/DBow3/DescManip.h
