+++ Lib/include/DBow3/QueryResults.h
