+++ Lib/include/DBow3/ScoringObject.h
