+++ Lib/include/DBow3/exports.h
@@ -35,17 +35,15 @@ or implied, of Rafael Muñoz Salinas.
 #define _CRT_SECURE_NO_DEPRECATE /* to avoid multiple Visual Studio 2005 warnings */
 #endif
 
-#if (defined WIN32 || defined _WIN32 || defined WINCE)   && defined DBOW_DSO_EXPORTS
-    #define DBOW_API __declspec(dllexport)
-    #pragma warning ( disable : 4251 ) //disable warning to templates with dll linkage.
-    #pragma warning ( disable : 4290 ) //disable warning due to exception specifications.
-    #pragma warning ( disable : 4996 ) //disable warning regarding unsafe vsprintf.
-    #pragma warning ( disable : 4244 ) //disable warning convesions with lost of data.
-
+#ifdef _WIN32
+#    ifdef DBOW_API
+#        define DBOW_API __declspec(dllexport)
+#    else
+#        define DBOW_API __declspec(dllimport)
+#    endif
 #else
-    #define DBOW_API
+#    define DBOW_API
 #endif
-
 
 #define  DBOW_VERSION "3.0.0"
 #endif
