+++ Lib/src/Database.cpp
@@ -1,4 +1,4 @@
-#include "Database.h"
+#include "DBow3/Database.h"
 
 namespace DBoW3{
